feat: 提供单独刷新截屏的接口
