+++ nestServer/.gitignore
@@ -1,7 +1,7 @@
 # compiled output
 /dist
 /node_modules
-
+/.cache
 # Logs
 logs
 *.log
