+++ nestServer/package.json
@@ -31,7 +31,7 @@
     "@nestjs/typeorm": "^9.0.1",
     "dayjs": "^1.11.6",
     "mysql2": "^2.3.3",
-    "puppeteer": "^19.3.0",
+    "puppeteer": "^19.5.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
